[feat]: classification functions depending on classifier argument
