+++ analyzer/serializers.py
@@ -1,36 +1,5 @@
 from rest_framework import serializers
-from .models import CaseStudy, ClassifyImageComponents, DecisionTreeTraining, FeatureExtractionTechnique, ExtractTrainingDataset, GUIComponentDetection
-
-class GUIComponentDetectionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = GUIComponentDetection
-        fields = '__all__' # ['eyetracking_log_filename', 'add_words_columns', 'overwrite_npy', 'algorithm']
-
-class ClassifyImageComponentsSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ClassifyImageComponents
-        fields = '__all__' # ['model_json_file_name', 'model_weights', 'model_properties', 'algorithm']
-
-class FeatureExtractionTechniqueSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = FeatureExtractionTechnique
-        fields = '__all__'
-
-class ExtractTrainingDatasetSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ExtractTrainingDataset
-        fields = '__all__' # ['columns_to_ignore']
-
-    def create(self, validated_data):
-        return ExtractTrainingDataset.objects.create(**validated_data)
-    
-class DecisionTreeTrainingSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = DecisionTreeTraining
-        fields = '__all__' # ['library', 'algorithms', 'mode', 'columns_to_ignore']
-
-    def create(self, validated_data):
-        return DecisionTreeTraining.objects.create(**validated_data)
+from .models import CaseStudy
 
 class CaseStudySerializer(serializers.ModelSerializer):
     phases_to_execute = serializers.JSONField()
