+++ featureextraction/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from . import views
 
 # urlpatterns = [
-#     path('classify-image-components/', views.classify_image_components.as_view(), name='classify-image-components'),
+#     path('classify-image-components/', views.ui_elements_classification.as_view(), name='classify-image-components'),
 #     path('detect-images-components/', views.detect_images_components.as_view(), name='detect-images-components'),
 #     path('ocr/', views.get_ocr_image.as_view(),name='ocr'),
 #     path('pad/', views.pad.as_view(), name='pad'),
